The sensitive detector for the straws has been moved from the traceback boxes to the individual straws themselves.
The traceback number in the StrawArtRecord was removed (as well as removed
from the example analysis code). More information will be put back into the
StrawArtRecord in future commits.

The Straws_service has been updated on how the row number, straw in row, and the station
number are extracted out of the Logical volume name when it's placed. Function created
which takes a string for an indicator and returns the number that follows it.