Updates to the straw hit and record to record the tracker number (based on location in the ring) that the particle hit.
however, the distribution of the tracker numbers is exhibiting some
funny behavior which needs further investigation. While each
straw is getting the correct traceback location, the tracks that hit the
strawSDs are only returning the first, or the last two trackers. And
it seems that each event returns these three locations within the event
(I would expect that each event would be in the same tracker, or at least multiple
hits in a row wihtin the same tracker). Don't know where the bug is but
wanted to save this work as is.