Updates to add a scallop extension on in only those scallop regions which there is a straw tracker system placed.
The buildUnionSolid function now takes two booleans in order
to know whether to add on extension to the first or second position
in the arc. It does this by looking at the StrawTrackerGeometry and
looking at the whichScallopLocations variables to find out where
the trackers are placed.

The variables for the size of the extension and the placement location
of the extensions have been placed in the VacuumGeomerty.