Updated the building of the support post to require that it only gets built if a boolean in the fcl file determines it should be.
All the componets of building the support post (the volume, the
place, the displaying) are now moved to be together within one if
statement.

At this time it isn't for sure if that post will be needed and thus
if it is the fcl parameter should be changed to reflect that.